Tests - remove user created in test_bind_address.
mtr is unhappy with unit.conc_connection otherwise, says
"MTR's internal check of the test case 'unit.conc_connection' failed."

+++ unittest/libmariadb/connection.c
@@ -841,6 +841,12 @@ static int test_bind_address(MYSQL *my)
     return SKIP;
   }
 
+  if (!bind_addr)
+  {
+    diag("No bind address specified");
+    return SKIP;
+  }
+
   sprintf(query, "DROP USER '%s'@'%s'", username, bind_addr);
   rc= mysql_query(my, query);
 
@@ -851,12 +857,6 @@ static int test_bind_address(MYSQL *my)
   sprintf(query, "GRANT ALL ON %s.* TO '%s'@'%s'", schema, username, bind_addr);
   rc= mysql_query(my, query);
   check_mysql_rc(rc, my);
-
-  if (!bind_addr)
-  {
-    diag("No bind address specified");
-    return SKIP;
-  }
 
   mysql= mysql_init(NULL);
   mysql_options(mysql, MYSQL_OPT_BIND, bind_addr);
@@ -870,6 +870,11 @@ static int test_bind_address(MYSQL *my)
   }
   diag("%s", mysql_get_host_info(mysql));
   mysql_close(mysql);
+
+  sprintf(query, "DROP USER '%s'@'%s'", username, bind_addr);
+  rc= mysql_query(my, query);
+  check_mysql_rc(rc, my);
+
   return OK;
 }
